ssize_t to int

+++ src/server/NetworkIOHandler.cpp
@@ -71,7 +71,7 @@ int NetworkIOHandler::receiveRequest( ConnectionManager& connManager, const int 
 	return 1;
 }
 
-ssize_t NetworkIOHandler::sendResponse( ConnectionManager &connManager, const int cli_sock )
+int NetworkIOHandler::sendResponse( ConnectionManager &connManager, const int cli_sock )
 {
 	std::vector<unsigned char> response = connManager.getFinalResponse( cli_sock );
 	size_t resSize = response.size();
